% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_agents}
\alias{available_agents}
\title{List Available RL Agents}
\usage{
available_agents()
}
\value{
A character vector of available RL agent names and their class types.
}
\description{
List Available RL Agents
}
\examples{

if (interactive()) available_agents()

}
