% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_agents_internal}
\alias{available_agents_internal}
\title{List Available RL Agents}
\usage{
available_agents_internal()
}
\value{
A data frame of available RL agent names and their class types. This assumes
the convention that all RL agents are prefaced with "agent_", e.g.,
"agent_tdrl_conditioning".
}
\description{
List Available RL Agents
}
