% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check_input.R
\name{check_action_reinforcement_episodes}
\alias{check_action_reinforcement_episodes}
\title{Check Action/Reinforcement Episodes}
\usage{
check_action_reinforcement_episodes(num_eps, action_ep, reinforcement_ep)
}
\arguments{
\item{num_eps}{Number of episodes per trial.}

\item{action_ep}{Episode number the action occurs on}

\item{reinforcement_ep}{Reinforcement number the action occurs on}
}
\value{
An error if the input is invalid, otherwise nothing.
}
\description{
Check Action/Reinforcement Episodes
}
\keyword{internal}
