% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check_input.R
\name{check_arm_input}
\alias{check_arm_input}
\title{Check Arm Input}
\usage{
check_arm_input(.input, num_arms, num_trials)
}
\arguments{
\item{.input}{The input passed into one of the array definition functions
created with \code{\link{rl_define_arms}}.}

\item{num_arms}{(numeric) The number of arms in the task.}

\item{num_trials}{(numeric) The number of trials to simulate.}
}
\value{
An error if the input is invalid, otherwise nothing.
}
\description{
Check Arm Input
}
\keyword{internal}
