% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_rl_simulate_action.R
\name{rl_action_simulate.greedy}
\alias{rl_action_simulate.greedy}
\title{Simulate an Action with a 'Greedy' Choice Policy}
\usage{
\method{rl_action_simulate}{greedy}(policy = "greedy", values, ...)
}
\arguments{
\item{policy}{Defines the action selection policy as "greedy"; argument
included in this method to support S3 Generics.}

\item{values}{A numeric vector containing the current value estimates of each
action.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A number representing which action will be taken.
}
\description{
This implementation of a 'greedy' action selection policy will
mean whichever action has the highest expected value will be taken.
}
\examples{

action <- numeric(100)
for (trial in seq_along(action)) {
 action[trial] <- rl_action_simulate(policy = "greedy",
                                    values = c(0.2, 0.25, 0.15, 0.8))
}

# All of the actions were to choose the highest value option
all(action == 4)

}
