% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_rl_simulate_action.R
\name{rl_action_simulate}
\alias{rl_action_simulate}
\title{Simulate an RL Agent's Action}
\usage{
rl_action_simulate(policy, values, ...)
}
\arguments{
\item{policy}{What policy should a decision be made under? Currently
supported are softmax, greedy, and epsilon-greedy.}

\item{values}{A numeric vector containing the current value estimates of each
action.}

\item{...}{Additional arguments passed to or from specific methods, such as
\code{tau} when \code{policy = "softmax"} and \code{epsilon} when \code{policy = "epsilonGreedy"}.}
}
\value{
A number representing which action will be taken given the chosen policy.
}
\description{
This is a generic function used to simulate an RL agent's action
given a specific decision-making policy.
}
