% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_arms.R
\name{rl_arms_define}
\alias{rl_arms_define}
\title{Define an RL Arm}
\usage{
rl_arms_define(probability, magnitude, alternative, trial)
}
\arguments{
\item{probability}{Either a numeric vector the same length as the number of
simulated trials or a single number between zero and one to define the
probability a reinforcement will be delivered.}

\item{magnitude}{Either a numeric vector the same length as the number of
simulated trials or a single number defining the magnitude of the
reinforcement to be delivered with probability \code{probability} if this arm is
chosen.}

\item{alternative}{Either a numeric vector the same length as the number of
simulated trials or a single number defining the magnitude of the
reinforcement to be delivered with probability \code{(1 - probability)} if this
arm is chosen.}

\item{trial}{A numeric vector describing the trial in which the probabilistic
reinforcement structure defined by the arguments \code{probability},
\code{magnitude}, and \code{alternative} applies.}
}
\value{
A data frame with columns 'probability', 'magnitude', 'alternative',
and 'trial' describing the reinforcement structure of an individual RL
Agent's arm.
}
\description{
Define the probabilistic-reinforcement structure of a specific
arm. This function is a simple wrapper around \code{data.frame}. As such, you
can pass in either numeric vectors the same length as the number of
simulated trials or a single number.
}
\examples{

#' # Define an arm that, when selected, leads to a 25\% probability of receiving a
# positive reinforcement with magnitude 1 or zero for the first 75 trials and a negative
# reinforcement with magnitude 1 or zero with 75\% probability for the second 75 trials.
rl_arms_define(
  probability = rep(0.25, 150),
  magnitude = c(rep(1, 75), rep(-1, 75)),
  alternative = 0,
  trial = 1:150
)

}
