% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_arms.R
\name{rl_arms_get_outcome}
\alias{rl_arms_get_outcome}
\title{Get Arm's Outcome based on its Probability and Reward Structure}
\usage{
rl_arms_get_outcome(arm_definitions, action, trial)
}
\arguments{
\item{arm_definitions}{A list of arm definitions where each element contains
a data frame with columns 'probability', 'magnitude', 'alternative', and
'trial' describing, respectively, the \code{probability} of receiving a reward
\code{magnitude} with the \code{alternative} for each \code{trial}.}

\item{action}{A numeric scalar representing which action was selected on a
given trial.}

\item{trial}{The trial in which an action was selected.}
}
\value{
A numeric reinforcement defined by \code{magnitude} (with probability
\code{prob}) or \code{alternative} (with probability \code{1 - prob}).
}
\description{
This function defines the reinforcement delivery for an
individual arm, and is used internaly by RL Bandit Agents. With probability
\code{prob}, it an arm will yield a reinforcement of \code{magnitude}; with
probability \code{1 - prob}, an arm will yield a reinforcement of \code{alternative}
(default of zero).
}
