% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_arms.R
\name{rl_arms_get_structure}
\alias{rl_arms_get_structure}
\title{Retrieve the List of Arm Definitions}
\usage{
rl_arms_get_structure(agent)
}
\arguments{
\item{agent}{An RL Agent created with \code{\link{rl_new_agent}}.}
}
\value{
If the RL Agent has its arm structure set, this will return the list
of arms defined with \code{\link{rl_set_arms}}; otherwise nothing.
}
\description{
Retrieve the List of Arm Definitions
}
