% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_arms.R
\name{rl_define_arms}
\alias{rl_define_arms}
\title{Define the Arm Structure}
\usage{
rl_define_arms(arm_input, num_arms, num_trials)
}
\arguments{
\item{arm_input}{A list of arm definitions where each element contains a data
frame with columns 'probability', 'magnitude', 'alternative', and 'trial'
describing, respectively, the \code{probability} of receiving a reward
\code{magnitude} with the \code{alternative} for each \code{trial}.}

\item{num_arms}{(numeric) The number of arms in the task.}

\item{num_trials}{(numeric) The number of trials to simulate.}
}
\description{
Define the Arm Structure
}
\details{
Each "arm" is a data frame with the following parameters:
\itemize{
\item \code{probability}: Either a numeric vector the same length as the number of
simulated trials or a single number between zero and one to define the
probability a reinforcement will be delivered.
\item \code{magnitude}: Either a numeric vector the same length as the number of
simulated trials or a single number defining the magnitude of the
reinforcement to be delivered with probability \code{probability} if this arm is
chosen.
\item \code{alternative}: Either a numeric vector the same length as the number of
simulated trials or a single number defining the magnitude of the
reinforcement to be delivered with probability \code{(1 - probability)} if this
arm is chosen.
\item \code{trial}: A numeric vector describing the trial in which the probabilistic
reinforcement structure defined by the arguments \code{probability},
\code{magnitude}, and \code{alternative} applies.
}
}
\examples{

# Define an arm that, when selected, leads to a 25\% probability of receiving
# positive reinforcement with magnitude 1 or zero for the first 75 trials and
# a negative reinforcement with magnitude 1 or zero with 75\% probability for
# the second 75 trials.

rl_define_arms(
  list(
    data.frame(
      probability = rep(0.25, 150),
      magnitude = c(rep(1, 75), rep(-1, 75)),
      alternative = 0,
      trial = 1:150
    )
  ),
  num_arms = 1,
  num_trials = 150
)
}
