% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-array_factory.R
\name{rl_define_array_factory}
\alias{rl_define_array_factory}
\title{Factory for Defining Array Setting Functions}
\usage{
rl_define_array_factory(type)
}
\arguments{
\item{type}{What type of array will be set? 'reinforcement' or 'cues'.}
}
\value{
A function to be exported for defining either array or cues.
}
\description{
Factory for Defining Array Setting Functions
}
\examples{

\dontrun{
rl_define_array_factory("reinforcement")
}


}
\keyword{internal}
