% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-array_factory.R
\name{rl_define_cues_array}
\alias{rl_define_cues_array}
\title{Define Cue Array for Tracking Cues in RL Agent's Environment}
\usage{
rl_define_cues_array(cue_input, num_episodes, num_trials)
}
\arguments{
\item{cue_input}{A list of of cues where each element contains a data frame
with columns 'onset', 'offset', 'magnitude', and 'trial' describing,
respectively, the episode number a cue presentation begins; the episode
number the cue presentation ends; the magnitude (saliency of a cue) of the
cue; the trials the cues occur.}

\item{num_episodes}{(numeric) The number of episodes per trial.}

\item{num_trials}{(numeric) The number of trials to simulate.}
}
\value{
A three-dimensional array tacking what cues are present in each
episode across all simulated trials.
}
\description{
Define Cue Array for Tracking Cues in RL Agent's Environment
}
\examples{

rl_define_cues_array(
  list(
    one = data.frame(
      onset = 3,
      offset = 8,
      magnitude = 1,
      trial = 1:100
    )
  ),
  num_episodes = 10,
  num_trials = 100
)

}
