% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_policies.R
\name{rl_define_policy}
\alias{rl_define_policy}
\title{Define the Action-Selection Policy for an RL Agent}
\usage{
rl_define_policy(policy, ...)
}
\arguments{
\item{policy}{What policy should a decision be made under? Currently
supported are softmax, greedy, and epsilon-greedy.}

\item{...}{Additional arguments passed to or from specific methods, such as
\code{tau} when \code{policy = "softmax"} and \code{epsilon} when \code{policy = "epsilonGreedy"}.}
}
\value{
The agent object with modified policy structure.
}
\description{
Define the Action-Selection Policy for an RL Agent
}
\examples{

# Example arguments include:

# (policy = "softmax", tau = 7)

# (policy = "epsilonGreedy", epsilon = 0.4)

# (policy = "greedy")
}
