% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_get_learned_values.R
\name{rl_get_learned_values}
\alias{rl_get_learned_values}
\title{Get RL Agent's Simulated Learned Values}
\usage{
rl_get_learned_values(agent, ...)
}
\arguments{
\item{agent}{An object of class "rlAgent" created with
\code{\link{rl_new_agent}}.}

\item{...}{Additional arguments to pass into the specific RL Agent for
recovering the learned value data.}
}
\value{
A cleaned version of the agent's learned values. See specific methods
for details.
}
\description{
This is a generic function for converting the agent's simulated
learned values into a nicer format, when applicable, for exploration and
visualization.
}
