% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_get_pe_data.R
\name{rl_get_pe_data.kArmedBandit}
\alias{rl_get_pe_data.kArmedBandit}
\title{Get 'kArmedBandit' Agent's Prediction Error Data}
\usage{
\method{rl_get_pe_data}{kArmedBandit}(agent, add_trial_zero = TRUE, trial_zero_value = 0, ...)
}
\arguments{
\item{agent}{An object of class "rlAgent" created with
\code{\link{rl_new_agent}}.}

\item{add_trial_zero}{(Logical) \code{TRUE} by default and trial zero will be
appended to the prediction error data frame with values from
\code{trial_zero_value}. \code{FALSE} and output will begin at trial one.}

\item{trial_zero_value}{(Numeric) Either a single value (default is 0) or a
vector of values with length \code{num_arms} times \code{num_episodes} to append for
trial 0.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A data frame with the agent's simulated prediction errors from taking
an action for each episode across all trials.
}
\description{
Convert the agent's prediction error array where, for each
action, there is an matrix with each row corresponding to an episode and
each column a trial with the experienced prediction error, to a data frame
with columns 'trial', 'episode', 'action', and 'value'.
}
