% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_get_pe_data.R
\name{rl_get_pe_data}
\alias{rl_get_pe_data}
\title{Get RL Agent's Simulated Prediction Errors}
\usage{
rl_get_pe_data(agent, ...)
}
\arguments{
\item{agent}{An object of class "rlAgent" created with
\code{\link{rl_new_agent}}.}

\item{...}{Additional arguments to pass into the specific RL Agent for
recovering the prediction error data.}
}
\value{
A cleaned version of the agent's prediction errors. See specific
methods for details.
}
\description{
This is a generic function for converting the agent's simulated
prediction errors into a nicer format, when applicable, for exploration and
visualization.
}
