% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_new_agent.R
\name{rl_new_agent}
\alias{rl_new_agent}
\title{Initialize a Reinforcement Learning Agent}
\usage{
rl_new_agent(model_type, model_id, ...)
}
\arguments{
\item{model_type}{What type of RL Agent would you like to create? One of
\code{\link{available_agents}}.}

\item{model_id}{model_id (character) A model identifier referencing the
reinforcement learning paradigm to perform the simulation.}

\item{...}{Additional arguments to pass into the specific RL Agent.
reinforcement learning paradigm to perform the simulation.}
}
\value{
A Reinforcement Learning Agent object of class \code{model_type}.
}
\description{
This is a generic function to create a new reinforcement
learning agent.
}
\examples{

rl_new_agent(model_type = "tdrlConditioning",
             model_id = "Example TDRL",
             num_trials = 100,
             num_episodes = 10,
             gamma = 1,
             alpha = 0.2)


}
