% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_new_agent.R
\name{rl_new_agent.tdrlConditioning}
\alias{rl_new_agent.tdrlConditioning}
\title{Initialize a 'tdrlConditioning' Agent}
\usage{
\method{rl_new_agent}{tdrlConditioning}(model_type, model_id, num_trials, num_episodes, gamma, alpha, ...)
}
\arguments{
\item{model_type}{What type of RL Agent would you like to create? One of
\code{\link{available_agents}}.}

\item{model_id}{model_id (character) A model identifier referencing the
reinforcement learning paradigm to perform the simulation.}

\item{num_trials}{(numeric) The number of trials to simulate.}

\item{num_episodes}{(numeric) The number of episodes per trial.}

\item{gamma}{(numeric) The temporal discounting factor of the RL agent}

\item{alpha}{(numeric) The learning rate of the RL agent}

\item{...}{Additional arguments passed to or from other methods.}
}
\description{
Create a new \code{tdrlConditioning} Agent
}
