% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_policies.R
\name{rl_policy_check_internal}
\alias{rl_policy_check_internal}
\title{Check Action-Selection Policy}
\usage{
rl_policy_check_internal(x)
}
\arguments{
\item{x}{A policy list defined by \code{\link{rl_set_policy_internal}}.}
}
\value{
If the policy is valid, nothing. If the policy definition is invalid, throws an error.
}
\description{
Check Action-Selection Policy
}
\keyword{internal}
