% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_arms.R
\name{rl_set_arms.kArmedBandit}
\alias{rl_set_arms.kArmedBandit}
\title{Set the Arm Structure for 'kArmedBandit' Agent}
\usage{
\method{rl_set_arms}{kArmedBandit}(agent, arm_input, keep_arm_structure = FALSE, ...)
}
\arguments{
\item{agent}{An RL Agent created with \code{\link{rl_new_agent}}.}

\item{arm_input}{A list of arm definitions where each element contains a data
frame with columns 'probability', 'magnitude', 'alternative', and 'trial'
describing, respectively, the \code{probability} of receiving a reward
\code{magnitude} with the \code{alternative} for each \code{trial}.}

\item{keep_arm_structure}{(Logical) \code{FALSE} (default) and any existing arm
structure will be replaced when called. \code{TRUE} and the arm structure will
be modified but will not remove previously defined rewards.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
The agent object with modified arm structure.
}
\description{
Set the Arm Structure for 'kArmedBandit' Agent
}
\examples{

# Annotated examples of different arm structures:

}
