% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_arms.R
\name{rl_set_arms}
\alias{rl_set_arms}
\title{Setup an agent's Arm Structure}
\usage{
rl_set_arms(agent, ...)
}
\arguments{
\item{agent}{An RL Agent created with \code{\link{rl_new_agent}}.}

\item{...}{Additional arguments to pass into the specific RL Agent for
setting the arm structure.}
}
\value{
The agent object with modified arm structure.
}
\description{
This is a generic function for, when applicable, defining the
arm structure of a reinforcement learning agent.
}
