% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_set_cues.R
\name{rl_set_cues}
\alias{rl_set_cues}
\title{Setup an agent's Cue Structure}
\usage{
rl_set_cues(agent, ...)
}
\arguments{
\item{agent}{An object of class "rlAgent" created with
\code{\link{rl_new_agent}}.}

\item{...}{Additional arguments to pass into the specific RL Agent for
setting the cue structure.}
}
\value{
The agent object with modified cue structure.
}
\description{
This is a generic function for, when applicable, defining the
cue structure of a reinforcement learning agent.
}
