% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_set_cues.R
\name{rl_set_cues.tdrlConditioning}
\alias{rl_set_cues.tdrlConditioning}
\title{Set the Cue Structure for 'tdrlConditioning' Agent}
\usage{
\method{rl_set_cues}{tdrlConditioning}(agent, cue_input, keep_cue_structure = FALSE, ...)
}
\arguments{
\item{agent}{An RL Agent created with \code{\link{rl_new_agent}}.}

\item{cue_input}{A list of of cues where each element contains
a data frame with columns 'onset', 'offset', 'magnitude', and 'trial'
describing, respectively, the episode number a cue presentation begins;
the episode number the cue presentation ends; the magnitude (saliency
of a cue) of the cue; the trials the cues occur.}

\item{keep_cue_structure}{(Logical) \code{FALSE} (default) and any existing
cue structure will be replaced when called. \code{TRUE} and the cue
structure will be modified but will not remove previously defined cues.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
The agent object with modified reward structure.
}
\description{
Define the onset episode and offset episode of cues for each
trial.
}
\examples{

# Annotated examples of different cue structures:

# For a task where one cue occurs during the first 250 trials with magnitude (1)
# and a second cue occurs during the second 250 trials with magnitude (1). Cue
# values are set to be zero if not explicitly defined.
ex1 <- list(
  one = data.frame(
    onset = 60,
    offset = 80,
    magnitude = c(rep(1, 250), rep(0, 250)),
    trial = 1:500),
  two = data.frame(
    onset = 70,
    offset = 80,
    magnitude = c(rep(0, 250), rep(1, 250)),
    trial = 1:500
  )
)

# This is an alternative way to specify the two-cue structure since the cue
# values are set to be zero if not explicitly defined.
ex2 <- list(
  one = data.frame(
    onset = 60,
    offset = 80,
    magnitude = 1,
    trial = 1:250),
  two = data.frame(
    onset = 70,
    offset = 80,
    magnitude = 1,
    trial = 251:500
  )
)



}
