% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_policies.R
\name{rl_set_policy.kArmedBandit}
\alias{rl_set_policy.kArmedBandit}
\title{Set the Action-Selection Policy for a 'kArmedBandit' Agent}
\usage{
\method{rl_set_policy}{kArmedBandit}(agent, policy, ...)
}
\arguments{
\item{agent}{An RL Agent created with \code{\link{rl_new_agent}}.}

\item{policy}{What policy should a decision be made under? Currently
supported are softmax, greedy, and epsilon-greedy.}

\item{...}{Additional arguments passed to or from specific methods, such as
\code{tau} when \code{policy = "softmax"} and \code{epsilon} when \code{policy = "epsilonGreedy"}.}
}
\value{
The agent object with modified policy structure.
}
\description{
Set the Action-Selection Policy for a 'kArmedBandit' Agent
}
\examples{

# Example arguments include:

# (policy = "softmax", tau = 7)

# (policy = "epsilonGreedy", epsilon = 0.4)

# (policy = "greedy")

}
