% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_policies.R
\name{rl_set_policy}
\alias{rl_set_policy}
\title{Setup an agent's Action-Selection Policy}
\usage{
rl_set_policy(agent, ...)
}
\arguments{
\item{agent}{An RL Agent created with \code{\link{rl_new_agent}}.}

\item{...}{Additional arguments to pass into the specific RL Agent for
setting the policy.}
}
\value{
The agent object with modified policy structure.
}
\description{
This is a generic function for, when applicable, defining the
action-selection policy of an RL Agent.
}
