% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_policies.R
\name{rl_set_policy_internal}
\alias{rl_set_policy_internal}
\title{Generate Policy List}
\usage{
rl_set_policy_internal(policy, ...)
}
\arguments{
\item{policy}{What policy should a decision be made under? Currently
supported are softmax, greedy, and epsilon-greedy.}

\item{...}{Additional arguments passed to or from specific methods, such as
\code{tau} when \code{policy = "softmax"} and \code{epsilon} when \code{policy = "epsilonGreedy"}.}
}
\value{
A list with a policy "type" and, if applicable, parameters.
}
\description{
Generate Policy List
}
\keyword{internal}
