% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_set_reinforcements.R
\name{rl_set_reinforcements}
\alias{rl_set_reinforcements}
\title{Setup an agent's Reward Structure}
\usage{
rl_set_reinforcements(agent, ...)
}
\arguments{
\item{agent}{An RL Agent created with \code{\link{rl_new_agent}}.}

\item{...}{Additional arguments to pass into the specific RL Agent for
setting the reward structure.}
}
\value{
The agent object with modified reward structure.
}
\description{
This is a generic function for, when applicable, defining the
reward structure of a reinforcement learning agent.
}
