% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_simulate_agent.R
\name{rl_simulate_agent.kArmedBandit}
\alias{rl_simulate_agent.kArmedBandit}
\title{Simulate the 'kArmedBandit' Agent}
\usage{
\method{rl_simulate_agent}{kArmedBandit}(agent, ...)
}
\arguments{
\item{agent}{An object of class "rlAgent" created with
\code{\link{rl_new_agent}}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
The agent object with the completed simulations.
}
\description{
Simulate the 'kArmedBandit' Agent
}
