% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_agent.R
\name{rl_simulate_agent}
\alias{rl_simulate_agent}
\title{Simulate the RL Agent}
\usage{
rl_simulate_agent(agent)
}
\arguments{
\item{agent}{An object of class "rlAgent" created with
\code{\link{rl_new_agent}}.}
}
\value{
The agent object with the completed simulations.
}
\description{
Simulate the RL Agent
}
