% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-rl_simulate_agent.R
\name{rl_simulate_agent.tdrlConditioning}
\alias{rl_simulate_agent.tdrlConditioning}
\title{Simulate the 'tdrlConditioning' Agent}
\usage{
\method{rl_simulate_agent}{tdrlConditioning}(agent, ...)
}
\arguments{
\item{agent}{An object of class "rlAgent" created with
\code{\link{rl_new_agent}}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
The agent object with the completed simulations.
}
\description{
Simulate the 'tdrlConditioning' Agent
}
