% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_prpt.R
\name{simulatePRPT}
\alias{simulatePRPT}
\title{Simulate PRPT task}
\usage{
simulatePRPT(num_iters = 1000, num_trials = 150, num_states = 4)
}
\arguments{
\item{num_iters}{Number of iterations}

\item{num_trials}{Number of trials}

\item{num_states}{Number of states -- should equal 4}
}
\value{
A list of the random earnings and correctness.
}
\description{
Simulate PRPT task
}
