% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_agent_template}
\alias{use_agent_template}
\title{Use an RL Agent Template from rlsims Package}
\usage{
use_agent_template(name)
}
\arguments{
\item{name}{Name of the agent template to use.}
}
\value{
Results of \link[base:readLines]{base::readLines} on the RL Agent template.
}
\description{
Use an RL Agent Template from rlsims Package
}
\examples{
\dontrun{
use_agent_template("tdrl_conditioning")
}
}
\keyword{internal}
