% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zeros}
\alias{zeros}
\title{Create a zero matrix.}
\usage{
zeros(dims)
}
\arguments{
\item{dims}{A numeric vector representing the dimensions of the array.}
}
\value{
A multi-dimensional array of zeros.
}
\description{
Inspired by MatLab's zeros function to easily create a
multi-dimensional array populated with zeros.
}
\examples{

# Create a 10-row by 2-column matrix of zeros
zeros(dims = c(10, 2))

# Create a 3-d array with 10-rows and 2-columns
zeros(dims = c(3, 10, 2))

}
