% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_agent.R
\name{rl_new_agent}
\alias{rl_new_agent}
\title{Initialize a Reinforcement Learning Agent}
\usage{
rl_new_agent(model_type, num_trials, num_episodes, num_cues, gamma, alpha)
}
\arguments{
\item{model_type}{(character) either "tdrl" or "vprl", referencing the
reinforcement learning paradigm to perform the simulation.}

\item{num_trials}{(numeric) The number of trials to simulate.}

\item{num_episodes}{(numeric) The number of episodes per trial.}

\item{num_cues}{(numeric) The number of potential cues interacting with the RL agent.}

\item{gamma}{(numeric) The temporal discounting factor of the RL agent}

\item{alpha}{(numeric) The learning rate of the RL agent}
}
\description{
Create a new \code{rlAgent} object
}
\examples{

mont6a <- rl_new_agent(model_type = "tdrl",
                       num_trials = 500,
                       num_episodes = 100,
                       num_cues = 2,
                       gamma = 1,
                       alpha = 0.05)

}
