% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_agent.R
\name{rl_set_cues}
\alias{rl_set_cues}
\title{Setup an agent's Cue Structure}
\usage{
rl_set_cues(agent, cue_list, keep_cue_structure = FALSE)
}
\arguments{
\item{agent}{An object of class "rlAgent" created with
\code{\link{rl_new_agent}}.}

\item{cue_list}{A list of length \code{num_cues} where each element contains a
data frame with columns 'cue', 'onset', 'offset', and 'trial'.}

\item{keep_cue_structure}{(Logical) \code{FALSE} (default) and any
existing cue structure will be replaced when called. \code{TRUE} and the
cue structure will be modified but will not remove previously
defined cues.}
}
\value{
The agent object with modified reward structure.
}
\description{
Define the onset episode and offset episode of cues for
each trial
}
\examples{
cue_data_list <- list(
  one = data.frame(
    cue = 1,
    onset = 60,
    offset = 80,
    trial = 1:500),
  two = data.frame(
    cue = 2,
    onset = 70,
    offset = 80,
    trial = 1:500
  )
)

}
